slash: disable stdio stream buffering
For normal operation, this is not strictly necessary with glibc which
automatically flushes the output stream when the input stream is read,
or for libc implementations that do not use buffering.

This does become necessary when the output is piped to another
application and the user wants to use slash interactively, since stdout
then becomes fully buffered.

Signed-off-by: Jeppe Ledet-Pedersen <jlp@satlab.com>