slash: rework argument handling and completion to reduce stack usage
Move the argv array into the slash context so the memory usage is known
at compile time. Also add a warning and return an error if too many
arguments are passed.

The completions array used to collect matching completion commands was
also allocated on the stack. Instead of also moving that to the slash
context, remove it completely and replace it with two iterations over
the command array: one to count the number of matches and find the
prefix length and one to print the completions, if needed.

Signed-off-by: Jeppe Ledet-Pedersen <jlp@satlab.com>