slash: optimize screen refresh
Erasing the line and rewriting everything on every change is simple, but
works poor with slow serial speeds and can still create visible refresh
artifacts with faster speeds.

Rewrite the refresh routine to keep track of the screen cursor and line
length and try to refresh fewer bytes. We still need a boolean to force
a full reset after we print completion lists. A full reset is also used
when browsing the history, which could probably be optimized further.

Signed-off-by: Jeppe Ledet-Pedersen <jlp@satlab.com>